requirements.txt
fastapi
uvicorn[standard]
oracledb
python-dotenv
mysql-connector-python
pydantic[email]